Implement advanced queue configuration
Allow for routing keys, exchanges, durable, exclusive, etc to be configured from
orc-agent's json config. This allows us to configure queues which ensure
exclusivity, or create consumers which are members of a "one-to-many" fanout
exchange, or share a similar routing key. Now we could listen on a broadcast
queue "all", which perhaps all orc-agent subscribers listen to, or specific
topics, like "production_webservers".

Wow, such possibility.